+++ Software/VNA_embedded/Application/Drivers/delay.hpp
@@ -9,6 +9,6 @@ void Init();
 uint64_t get_us();
 
 void ms(uint32_t t);
-void us(uint32_t t);
+void us(uint16_t t);
 
 }
